Use unsigned short for lat/lon

+++ gensphere.h
@@ -54,11 +54,11 @@ public:
   }
 };
 
-void gensphere (geom_t *, int *, float **, 
+void gensphere (geom_t *, int *, unsigned short **, 
                 unsigned int *, float **, const std::string &);
 
 void gensphere_grib 
-               (geom_t *, int *, float **, 
+               (geom_t *, int *, unsigned short **, 
                 unsigned int *, float **, const std::string &);
